149 dificukty ratings in a table (#150)
* corrected procent calcaution

* create diffuclty bar

* adjuest diffculty message senfing

* formatting js

+++ ProjectSource/RedoCode_backend/src/main/java/com/redocode/backend/database/ExcersizeListEntry.java
@@ -15,7 +15,7 @@ public class ExcersizeListEntry {
   private Long id;
   private String name;
   private Set<ProgrammingLanguage> languages;
-  private int difficulty;
+  private float difficulty;
   private Long popularity;
   private String description;
 
@@ -26,9 +26,9 @@ public class ExcersizeListEntry {
     this.languages = e.getLanguages();
     this.difficulty =
         e.getRatings().size() != 0
-            ? e.getRatings().stream().mapToInt(element -> element.getRating()).sum()
+            ? (float)e.getRatings().stream().mapToDouble(element -> element.getRating()).sum()
                 / e.getRatings().size()
-            : 0;
+            : 0f;
     this.popularity = (long) e.getSolutions().size();
     this.description = e.getDescription();
   }
